reshape lib folder + add utils folder + add constants.ts
