+++ package.json
@@ -53,6 +53,7 @@
     "react-day-picker": "8.10.1",
     "react-dom": "^19.0.0",
     "recharts": "^2.15.2",
+    "server-only": "^0.0.1",
     "tailwind-merge": "^3.2.0",
     "tailwind-variants": "^1.0.0",
     "zod": "^3.24.2"
