+++ src/components/expenses/pending-expenses.tsx
@@ -1,8 +1,7 @@
-import { getUser } from "@/lib/utils";
+import { getUser, getCurrency } from "@/lib/utils/server-only-utils";
 import { getExpenses, getTotalAmountExpenses } from "@/lib/queries/expense";
 import { redirect } from "next/navigation";
 import PendingExpensesTable from "./pending-expenses-table";
-import { getCurrency } from "@/lib/cookies";
 import { formatPriceWithCurrency } from "@/lib/utils";
 
 const pendingExpenses = {
