+++ src/components/tremor-raw/inputs/input.tsx
@@ -3,8 +3,12 @@
 import React from "react";
 import { RiEyeFill, RiEyeOffFill, RiSearchLine } from "@remixicon/react";
 import { tv, type VariantProps } from "tailwind-variants";
-
-import { cx, focusInput, focusRing, hasErrorInput } from "@/lib/utils";
+import {
+  cx,
+  focusInput,
+  focusRing,
+  hasErrorInput,
+} from "@/lib/utils/tremor-raw/utils";
 
 const inputStyles = tv({
   base: [
