+++ src/components/tremor-raw/inputs/label.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import * as LabelPrimitives from "@radix-ui/react-label";
 
-import { cx } from "@/lib/utils";
+import { cx } from "@/lib/utils/tremor-raw/utils";
 
 interface LabelProps
   extends React.ComponentPropsWithoutRef<typeof LabelPrimitives.Root> {
