+++ src/components/tremor-raw/inputs/select.tsx
@@ -9,7 +9,7 @@ import {
   RiExpandUpDownLine,
 } from "@remixicon/react";
 
-import { cx, focusInput, hasErrorInput } from "@/lib/utils";
+import { cx, focusInput, hasErrorInput } from "@/lib/utils/tremor-raw/utils";
 
 const Select = SelectPrimitives.Root;
 Select.displayName = "Select";
