+++ src/components/tremor-raw/ui/badge.tsx
@@ -3,7 +3,7 @@
 import React from "react";
 import { tv, type VariantProps } from "tailwind-variants";
 
-import { cx } from "@/lib/utils";
+import { cx } from "@/lib/utils/tremor-raw/utils";
 
 const badgeVariants = tv({
   base: cx(
