+++ src/components/tremor-raw/ui/dialog.tsx
@@ -2,8 +2,7 @@
 
 import React from "react";
 import * as DialogPrimitives from "@radix-ui/react-dialog";
-
-import { cx, focusRing } from "@/lib/utils";
+import { cx, focusRing } from "@/lib/utils/tremor-raw/utils";
 
 const Dialog = (
   props: React.ComponentPropsWithoutRef<typeof DialogPrimitives.Root>
