+++ src/components/tremor-raw/ui/drawer.tsx
@@ -3,7 +3,7 @@
 import * as React from "react";
 import * as DrawerPrimitives from "@radix-ui/react-dialog";
 import { RiCloseLine } from "@remixicon/react";
-import { cx, focusRing } from "@/lib/utils";
+import { cx, focusRing } from "@/lib/utils/tremor-raw/utils";
 import { Button } from "./button";
 
 const Drawer = (
