+++ src/components/tremor-raw/ui/table.tsx
@@ -2,7 +2,7 @@
 
 import React from "react";
 
-import { cx } from "@/lib/utils";
+import { cx } from "@/lib/utils/tremor-raw/utils";
 
 const TableRoot = React.forwardRef<
   HTMLDivElement,
