--- src/lib/cookies.ts
@@ -1,6 +1,0 @@
-import { cookies } from "next/headers";
-
-export async function getCurrency() {
-  const cookiesStore = await cookies();
-  return cookiesStore.get("currency")?.value;
-}
